Improves styling and componentizes new content
