+++ postcss.config.js
@@ -1,6 +1,6 @@
 module.exports = {
-  plugins: [
-    require('tailwindcss'),
-    require('postcss-preset-env'),
-  ],
-};
+  plugins: {
+    tailwindcss: {},
+    autoprefixer: {},
+  },
+}
