--- src/assets/styles/tailwind.css
@@ -1,30 +1,0 @@
-@tailwind base;
-
-/* Your own custom base styles */
-
-/* Start purging... */
-@tailwind components;
-/* Stop purging. */
-
-/* Your own custom component styles */
-
-/* Start purging... */
-@tailwind utilities;
-/* Stop purging. */
-
-/* Your own custom utilities */
-
-@font-face {
-  font-family: "BandeinsSans";
-  src: local("BandeinsSans"),
-    url(../fonts/BandeinsSansRegular.otf) format("opentype");
-}
-
-body {
-  background-color: black;
-  background-image: url(../images/background.jpg);
-  background-position: center center;
-  background-repeat: no-repeat;
-  background-size: cover;
-  background-attachment: fixed;
-}
