+++ src/components/ContentBlock.tsx
@@ -1,0 +1,3 @@
+export const ContentBlock: React.FC = ({ children }) => (
+  <div className="backdrop-blur rounded-3xl md:p-10 p-5 my-4 md:flex-none flex-1 shadow-xl shadow-black">{children}</div>
+)
