+++ src/content/convert.json
@@ -1,0 +1,4 @@
+{
+  "header": "Redeem PHONON",
+  "paragraphs": ["Convert your GRID to PHONON"]
+}
